+++ SplitListWebApi/ApiFormat/ShadowTables/RecipeItem.cs
@@ -5,10 +5,10 @@ namespace ApiFormat.ShadowTables
 {
     public class RecipeItem
     {
-        public int ItemID { get; set; }
+        public double ItemID { get; set; }
         public ItemModel item { get; set; }
 
-        public int RecipeID { get; set; }
+        public double RecipeID { get; set; }
         public RecipeModel Recipe { get; set; }
     }
 }
