+++ SplitListWebApi/ApiFormat/ShadowTables/UserGroup.cs
@@ -4,9 +4,9 @@ namespace ApiFormat.ShadowTables
 {
     public class UserGroup
     {
-        public int UserID { get; set; }
+        public double UserID { get; set; }
         public UserModel User { get; set; }
-        public int GroupID { get; set; }
+        public double GroupID { get; set; }
         public GroupModel Group { get; set; }
     }
 }
